source install trick
Install script will now pop you into a venv post script which makes the user setup so much easier.

+++ Server/install.sh
@@ -1,15 +1,44 @@
 #!/bin/bash
 
+# make sure source calls this script
+if [[ "${BASH_SOURCE[0]}" == "${0}" ]]; then
+    echo "Error: This script must be run using 'source', ex 'source install.sh'"
+    echo "Usage: source $0"
+    exit 1
+fi
+
 #sudo apt-get install pip -y
 
-echo "Creating virtual environment 'server_venv'..."
-python3 -m venv server_venv
+
+# Check if venv is available
+if ! python3 -m venv --help > /dev/null 2>&1; then
+    echo "Error: 'venv' module is not available. Please ensure Python 3 is installed."
+    exit 1
+fi
+
+
+# Define the virtual environment name
+VENV_NAME="server_venv"
+
+# Create the virtual environment if it doesn't exist
+if [ ! -d "$VENV_NAME" ]; then
+    echo "Creating virtual environment '$VENV_NAME'..."
+    python3 -m venv "$VENV_NAME"
+else
+    echo "Virtual environment '$VENV_NAME' already exists."
+fi
 
 # Activate the virtual environment
-source server_venv/bin/activate
+source "$VENV_NAME/bin/activate"
 
-echo "installing pip reqs"
-pip install -r requirements.txt
+# Check if requirements.txt exists
+if [ -f "requirements.txt" ]; then
+    echo "Installing pip requirements..."
+    pip install -r requirements.txt
+else
+    echo "Error: requirements.txt not found."
+    exit 1
+fi
 
 # Ensure yq is installed
 if ! command -v yq &> /dev/null
@@ -48,9 +77,31 @@ then
         exit 1
     fi
 fi
+## make sure docker is installed
+if ! command -v docker &> /dev/null
+then
+    echo "start.sh: docker could not be found."
+
+    # Prompt the user for installation
+    read -p "start.sh: Do you want to install yq locally at ~/.local/bin/yq? [Y/n]: " answer
+    answer=${answer:-Y}  # Default to 'Y' if no input is provided
+
+    if [[ "$answer" =~ ^[Yy]$ ]]; then
+        echo "start.sh: Installing docker..."
+
+        sudo apt-get install docker.io
+    else
+        echo "Installation aborted. Cannot continue without docker."
+        exit 1
+    fi
+fi
+
 
 echo "Installing GUnicorn"
 sudo apt-get install gunicorn
 
 echo "Pulling & starting redis docker container"
+#may need to pull this based on docker in python implementation
 sudo docker run -d --name redis-stack-server -p 6379:6379 redis/redis-stack-server:latest 
+
+echo "SUCCESS... hopefully"
